Refactoring to make PluginDescriptor more usable
