+++ pf4j/src/test/java/org/pf4j/DefaultPluginFactoryTest.java
@@ -38,7 +38,7 @@ public class DefaultPluginFactoryTest {
      */
     @Test
     public void testCreate() {
-        PluginDescriptor pluginDescriptor = mock(PluginDescriptor.class);
+        PluginDescriptor pluginDescriptor = mock(DefaultPluginDescriptor.class);
         when(pluginDescriptor.getPluginClass()).thenReturn(TestPlugin.class.getName());
 
         PluginWrapper pluginWrapper = mock(PluginWrapper.class);
@@ -57,7 +57,7 @@ public class DefaultPluginFactoryTest {
      */
     @Test
     public void testCreateFail() {
-        PluginDescriptor pluginDescriptor = mock(PluginDescriptor.class);
+        PluginDescriptor pluginDescriptor = mock(DefaultPluginDescriptor.class);
         when(pluginDescriptor.getPluginClass()).thenReturn(FailTestPlugin.class.getName());
 
         PluginWrapper pluginWrapper = mock(PluginWrapper.class);
@@ -75,7 +75,7 @@ public class DefaultPluginFactoryTest {
      */
     @Test
     public void testCreateFailNotFound() {
-        PluginDescriptor pluginDescriptor = mock(PluginDescriptor.class);
+        PluginDescriptor pluginDescriptor = mock(DefaultPluginDescriptor.class);
         when(pluginDescriptor.getPluginClass()).thenReturn("org.pf4j.plugin.NotFoundTestPlugin");
 
         PluginWrapper pluginWrapper = mock(PluginWrapper.class);
@@ -93,7 +93,7 @@ public class DefaultPluginFactoryTest {
      */
     @Test
     public void testCreateFailConstructor() {
-        PluginDescriptor pluginDescriptor = mock(PluginDescriptor.class);
+        PluginDescriptor pluginDescriptor = mock(DefaultPluginDescriptor.class);
         when(pluginDescriptor.getPluginClass()).thenReturn(AnotherFailTestPlugin.class.getName());
 
         PluginWrapper pluginWrapper = mock(PluginWrapper.class);
